chore: dividing layers of chat screen to follow MVVM archicture
